requirements.txt
selenium~=4.34.2
pytest~=8.4.1
pytest-rerunfailures
allure-pytest
allure-python-commons~=2.15.0